# backend-projects-js/12-url-shortening-service/server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: SnapRoute Engine | Complete API Documentation
  version: 1.0.0
  description: Fully comprehensive API specification mapping every route of the high-performance URL shortener backend.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
paths:
  /auth/register:
    post:
      summary: Register a new account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  minLength: 3
                  example: darthCoder
                email:
                  type: string
                  format: email
                  example: darth@example.com
                password:
                  type: string
                  format: password
                  example: StrongPassword123!
      responses:
        '201':
          description: User created successfully
        '400':
          $ref: '#/components/responses/ErrorResponse'

  /auth/login:
    post:
      summary: Authenticate user & issue JWT
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: darth@example.com
                password:
                  type: string
                  format: password
                  example: StrongPassword123!
      responses:
        '200':
          description: Authentication successful
        '401':
          $ref: '#/components/responses/ErrorResponse'

  /urls:
    post:
      summary: Create a shortened URL
      tags:
        - URLs
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - originalUrl
              properties:
                originalUrl:
                  type: string
                  example: https://roadmap.sh/projects/url-shortening-service
                customAlias:
                  type: string
                  example: my-custom-link
                password:
                  type: string
                  example: secret123
                expiresAt:
                  type: string
                  format: date-time
                tags:
                  type: array
                  items:
                    type: string
                  example: ["dev", "roadmap"]
                isFavorite:
                  type: boolean
                  example: true
      responses:
        '201':
          description: URL shortened successfully
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'

    get:
      summary: Get user URL portfolio dashboard with filters
      tags:
        - URLs
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search term matching originalUrl or shortCode
        - in: query
          name: tag
          schema:
            type: string
          description: Filter by specific tag
        - in: query
          name: isFavorite
          schema:
            type: boolean
          description: Filter favorite links
        - in: query
          name: page
          schema:
            type: integer
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
          description: Number of items per page
      responses:
        '200':
          description: Dashboard portfolio data retrieved successfully
        '401':
          $ref: '#/components/responses/ErrorResponse'

    delete:
      summary: Bulk delete multiple URLs
      tags:
        - URLs
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ids
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  example: ["64a1b2c3d4e5f6g7h8i9j0k1", "64a1b2c3d4e5f6g7h8i9j0k2"]
      responses:
        '200':
          description: Batch deleted successfully
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'

  /urls/{id}:
    patch:
      summary: Update a URL asset configuration
      tags:
        - URLs
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
                isFavorite:
                  type: boolean
                password:
                  type: string
                  nullable: true
                expiresAt:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        '200':
          description: URL updated and Redis cache invalidated successfully
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

    delete:
      summary: Permanently delete a single URL asset
      tags:
        - URLs
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: URL permanently deleted and evicted from cache
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /urls/{id}/archive:
    patch:
      summary: Toggle archive state for a URL
      tags:
        - URLs
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - isArchived
              properties:
                isArchived:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Archive status updated successfully
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /urls/check-alias/{alias}:
    get:
      summary: Check if a custom alias is available
      tags:
        - URLs
      parameters:
        - in: path
          name: alias
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns availability boolean status
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                    example: true

  /urls/analytics/{shortCode}:
    get:
      summary: Get detailed click analytics breakdown
      tags:
        - URLs
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: shortCode
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Click metrics metadata retrieved successfully
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /health:
    get:
      summary: Service health check
      tags:
        - System
      responses:
        '200':
          description: Services operational status (MongoDB & Redis UP)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
                  services:
                    type: object
                    properties:
                      database:
                        type: string
                        example: UP
                      cache:
                        type: string
                        example: UP

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Standard error response layout
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Error description message.